+++ tests/testthat/test-paramset.R
@@ -907,3 +907,19 @@ test_that("checks work for cond inc/dec", {
   p$values$a <- 3:1
   expect_error(p$values$a <- c(3, 3, 2), "not strictly decreasing")
 })
+
+test_that("checks multiple conditions can work/fail", {
+  p <- pset(
+    prm("a", "nreals"),
+    prm("b", "nreals"),
+    deps = list(
+      list(id = "a", cond = cnd("inc")),
+      list(id = "a", on = "b", cond = cnd("len", id = "b"))
+    )
+  )
+  expect_error(p$values$a <- 3:1, "not increasing")
+  expect_error(p$values <- list(a = 1, b = 1:2), "len")
+  expect_error(p$values <- list(a = 3:1, b = 1:3), "increasing")
+  p$values <- list(a = 1:3, b = 1:3)
+  expect_error(p$values <- list(a = 1:2, b = 1:3), "len")
+})